Split the InBandKeyManager into InBand(Encryption|Decryption)Manager
