+++ kroxylicious-filters/kroxylicious-encryption/src/test/java/io/kroxylicious/filter/encryption/inband/InBandDecryptionManagerTest.java
@@ -62,7 +62,7 @@ import static org.mockito.Mockito.doAnswer;
 import static org.mockito.Mockito.doReturn;
 import static org.mockito.Mockito.when;
 
-class InBandKeyManagerTest {
+class InBandDecryptionManagerTest {
 
     private static final String ARBITRARY_KEY = "key";
     private static final String ARBITRARY_KEY_2 = "key2";
@@ -86,7 +86,8 @@ class InBandKeyManagerTest {
     @Test
     void shouldEncryptRecordValue() {
         InMemoryKms kms = getInMemoryKms();
-        var km = createKeyManager(kms, 500_000);
+        var encryptionManager = createEncryptionManager(kms, 500_000);
+        var decryptionManager = createDecryptionManager(kms);
 
         var kekId = kms.generateKey();
 
@@ -95,7 +96,7 @@ class InBandKeyManagerTest {
 
         List<Record> encrypted = new ArrayList<>();
         List<Record> initial = List.of(record);
-        assertThat(doEncrypt(km, "topic", 1, new EncryptionScheme<>(kekId, EnumSet.of(RecordField.RECORD_VALUE)), initial, encrypted))
+        assertThat(doEncrypt(encryptionManager, "topic", 1, new EncryptionScheme<>(kekId, EnumSet.of(RecordField.RECORD_VALUE)), initial, encrypted))
                 .isCompleted();
         assertThat(encrypted.iterator())
                 .toIterable()
@@ -106,7 +107,7 @@ class InBandKeyManagerTest {
         List<Record> decrypted = new ArrayList<>();
         String topic = "foo";
         int partition = 1;
-        assertThat(doDecrypt(km, topic, partition, encrypted, decrypted))
+        assertThat(doDecrypt(decryptionManager, topic, partition, encrypted, decrypted))
                 .isCompleted();
 
         assertThat(decrypted.iterator())
@@ -121,7 +122,7 @@ class InBandKeyManagerTest {
         // given
         InMemoryKms kms = getInMemoryKms();
         EncryptionScheme<UUID> scheme = createScheme(kms);
-        var km = createKeyManager(kms, 500_000);
+        var encryptionManager = createEncryptionManager(kms, 500_000);
 
         MutableRecordBatch firstBatch = RecordTestUtils.singleElementRecordBatch(RecordBatch.CURRENT_MAGIC_VALUE, 1L, CompressionType.GZIP, TimestampType.CREATE_TIME, 2L,
                 3L,
@@ -137,7 +138,7 @@ class InBandKeyManagerTest {
         MemoryRecords records = RecordTestUtils.memoryRecords(firstBatch, secondBatch);
 
         // when
-        MemoryRecords encrypted = assertImmediateSuccessAndGet(encrypt(km, scheme, records));
+        MemoryRecords encrypted = assertImmediateSuccessAndGet(encrypt(encryptionManager, scheme, records));
 
         // then
         MemoryRecordsAssert encryptedAssert = MemoryRecordsAssert.assertThat(encrypted);
@@ -151,7 +152,8 @@ class InBandKeyManagerTest {
         // given
         InMemoryKms kms = getInMemoryKms();
         EncryptionScheme<UUID> scheme = createScheme(kms);
-        var km = createKeyManager(kms, 500_000);
+        var encryptionManager = createEncryptionManager(kms, 500_000);
+        var decryptionManager = createDecryptionManager(kms);
 
         MutableRecordBatch firstBatch = RecordTestUtils.singleElementRecordBatch(RecordBatch.CURRENT_MAGIC_VALUE, 1L, CompressionType.GZIP, TimestampType.CREATE_TIME, 2L,
                 3L,
@@ -165,10 +167,10 @@ class InBandKeyManagerTest {
                         StandardCharsets.UTF_8),
                 ARBITRARY_VALUE_2.getBytes(StandardCharsets.UTF_8));
         MemoryRecords records = RecordTestUtils.memoryRecords(firstBatch, secondBatch);
-        MemoryRecords encrypted = assertImmediateSuccessAndGet(encrypt(km, scheme, records));
+        MemoryRecords encrypted = assertImmediateSuccessAndGet(encrypt(encryptionManager, scheme, records));
 
         // when
-        MemoryRecords decrypted = assertImmediateSuccessAndGet(decrypt(km, encrypted));
+        MemoryRecords decrypted = assertImmediateSuccessAndGet(decrypt(decryptionManager, encrypted));
 
         // then
         MemoryRecordsAssert decryptedAssert = MemoryRecordsAssert.assertThat(decrypted);
@@ -182,7 +184,7 @@ class InBandKeyManagerTest {
         // given
         InMemoryKms kms = getInMemoryKms();
         EncryptionScheme<UUID> scheme = createScheme(kms);
-        var km = createKeyManager(kms, 500_000);
+        var encryptionManager = createEncryptionManager(kms, 500_000);
 
         MutableRecordBatch firstBatch = RecordTestUtils.singleElementRecordBatch(1L, ARBITRARY_KEY, ARBITRARY_VALUE, ABSENT_HEADERS);
         MutableRecordBatch controlBatch = RecordTestUtils.abortTransactionControlBatch(2);
@@ -190,7 +192,7 @@ class InBandKeyManagerTest {
         MemoryRecords records = RecordTestUtils.memoryRecords(firstBatch, controlBatch);
 
         // when
-        MemoryRecords encrypted = assertImmediateSuccessAndGet(encrypt(km, scheme, records));
+        MemoryRecords encrypted = assertImmediateSuccessAndGet(encrypt(encryptionManager, scheme, records));
 
         // then
         MemoryRecordsAssert encryptedAssert = MemoryRecordsAssert.assertThat(encrypted);
@@ -204,16 +206,17 @@ class InBandKeyManagerTest {
         // given
         InMemoryKms kms = getInMemoryKms();
         EncryptionScheme<UUID> scheme = createScheme(kms);
-        var km = createKeyManager(kms, 500_000);
+        var encryptionManager = createEncryptionManager(kms, 500_000);
+        var decryptionManager = createDecryptionManager(kms);
 
         MutableRecordBatch firstBatch = RecordTestUtils.singleElementRecordBatch(1L, ARBITRARY_KEY, ARBITRARY_VALUE, ABSENT_HEADERS);
         MutableRecordBatch controlBatch = RecordTestUtils.abortTransactionControlBatch(2);
         Record controlRecord = controlBatch.iterator().next();
         MemoryRecords records = RecordTestUtils.memoryRecords(firstBatch, controlBatch);
-        MemoryRecords encrypted = assertImmediateSuccessAndGet(encrypt(km, scheme, records));
+        MemoryRecords encrypted = assertImmediateSuccessAndGet(encrypt(encryptionManager, scheme, records));
 
         // when
-        MemoryRecords decrypted = assertImmediateSuccessAndGet(decrypt(km, encrypted));
+        MemoryRecords decrypted = assertImmediateSuccessAndGet(decrypt(decryptionManager, encrypted));
 
         // then
         MemoryRecordsAssert decryptedAssert = MemoryRecordsAssert.assertThat(decrypted);
@@ -227,14 +230,14 @@ class InBandKeyManagerTest {
         // given
         InMemoryKms kms = getInMemoryKms();
         EncryptionScheme<UUID> scheme = createScheme(kms);
-        var km = createKeyManager(kms, 500_000);
+        var encryptionManager = createEncryptionManager(kms, 500_000);
 
         MutableRecordBatch firstBatch = RecordTestUtils.singleElementRecordBatch(1L, ARBITRARY_KEY, ARBITRARY_VALUE, ABSENT_HEADERS);
         MutableRecordBatch emptyBatch = RecordTestUtils.recordBatchWithAllRecordsRemoved(2L);
         MemoryRecords records = RecordTestUtils.memoryRecords(firstBatch, emptyBatch);
 
         // when
-        MemoryRecords encrypted = assertImmediateSuccessAndGet(encrypt(km, scheme, records));
+        MemoryRecords encrypted = assertImmediateSuccessAndGet(encrypt(encryptionManager, scheme, records));
 
         // then
         MemoryRecordsAssert encryptedAssert = MemoryRecordsAssert.assertThat(encrypted);
@@ -248,15 +251,16 @@ class InBandKeyManagerTest {
         // given
         InMemoryKms kms = getInMemoryKms();
         EncryptionScheme<UUID> scheme = createScheme(kms);
-        var km = createKeyManager(kms, 500_000);
+        var encryptionManager = createEncryptionManager(kms, 500_000);
+        var decryptionManager = createDecryptionManager(kms);
 
         MutableRecordBatch firstBatch = RecordTestUtils.singleElementRecordBatch(1L, ARBITRARY_KEY, ARBITRARY_VALUE, ABSENT_HEADERS);
         MutableRecordBatch emptyBatch = RecordTestUtils.recordBatchWithAllRecordsRemoved(2L);
         MemoryRecords records = RecordTestUtils.memoryRecords(firstBatch, emptyBatch);
-        MemoryRecords encrypted = assertImmediateSuccessAndGet(encrypt(km, scheme, records));
+        MemoryRecords encrypted = assertImmediateSuccessAndGet(encrypt(encryptionManager, scheme, records));
 
         // when
-        MemoryRecords decrypted = assertImmediateSuccessAndGet(decrypt(km, encrypted));
+        MemoryRecords decrypted = assertImmediateSuccessAndGet(decrypt(decryptionManager, encrypted));
 
         // then
         MemoryRecordsAssert decryptedAssert = MemoryRecordsAssert.assertThat(decrypted);
@@ -266,14 +270,14 @@ class InBandKeyManagerTest {
     }
 
     @NonNull
-    private static CompletionStage<Void> doDecrypt(InBandKeyManager<UUID, InMemoryEdek> km, String topic, int partition, List<Record> encrypted,
+    private static CompletionStage<Void> doDecrypt(InBandDecryptionManager<UUID, InMemoryEdek> km, String topic, int partition, List<Record> encrypted,
                                                    List<Record> decrypted) {
         return km.decrypt(topic, partition, RecordTestUtils.memoryRecords(encrypted), ByteBufferOutputStream::new)
                 .thenAccept(records -> records.records().forEach(decrypted::add));
     }
 
     @NonNull
-    private static CompletionStage<Void> doEncrypt(InBandKeyManager<UUID, InMemoryEdek> keyManager,
+    private static CompletionStage<Void> doEncrypt(InBandEncryptionManager<UUID, InMemoryEdek> keyManager,
                                                    String topic,
                                                    int partition,
                                                    EncryptionScheme<UUID> scheme,
@@ -289,7 +293,8 @@ class InBandKeyManagerTest {
     @Test
     void shouldTolerateEncryptingAndDecryptingEmptyRecordValue() {
         InMemoryKms kms = getInMemoryKms();
-        var km = createKeyManager(kms, 500_000);
+        var encryptionManager = createEncryptionManager(kms, 500_000);
+        var decryptionManager = createDecryptionManager(kms);
 
         var kekId = kms.generateKey();
 
@@ -298,14 +303,14 @@ class InBandKeyManagerTest {
 
         List<Record> encrypted = new ArrayList<>();
         List<Record> initial = List.of(record);
-        assertThat(doEncrypt(km, "topic", 1, new EncryptionScheme<>(kekId, EnumSet.of(RecordField.RECORD_VALUE)), initial, encrypted))
+        assertThat(doEncrypt(encryptionManager, "topic", 1, new EncryptionScheme<>(kekId, EnumSet.of(RecordField.RECORD_VALUE)), initial, encrypted))
                 .isCompleted();
         record.value().rewind();
         assertEquals(1, encrypted.size());
         assertNotEquals(initial, encrypted);
 
         List<Record> decrypted = new ArrayList<>();
-        assertThat(doDecrypt(km, "foo", 1, encrypted, decrypted)).isCompleted();
+        assertThat(doDecrypt(decryptionManager, "foo", 1, encrypted, decrypted)).isCompleted();
 
         assertEquals(initial, decrypted);
     }
@@ -313,13 +318,13 @@ class InBandKeyManagerTest {
     @Test
     void decryptSupportsUnencryptedRecordValue() {
         InMemoryKms kms = getInMemoryKms();
-        var km = createKeyManager(kms, 500_000);
+        var decryptionManager = createDecryptionManager(kms);
 
         byte[] recBytes = { 1, 2, 3 };
         Record record = RecordTestUtils.record(recBytes);
 
         List<Record> received = new ArrayList<>();
-        assertThat(doDecrypt(km, "foo", 1, List.of(record), received)).isCompleted();
+        assertThat(doDecrypt(decryptionManager, "foo", 1, List.of(record), received)).isCompleted();
 
         assertThat(received).hasSize(1);
         assertThat(received.stream()
@@ -335,8 +340,8 @@ class InBandKeyManagerTest {
     @MethodSource
     void decryptSupportsEmptyRecordBatches(MemoryRecords records) {
         InMemoryKms kms = getInMemoryKms();
-        var km = createKeyManager(kms, 500_000);
-        assertThat(km.decrypt("foo", 1, records, ByteBufferOutputStream::new))
+        var decryptionManager = createDecryptionManager(kms);
+        assertThat(decryptionManager.decrypt("foo", 1, records, ByteBufferOutputStream::new))
                 .succeedsWithin(Duration.ZERO).isSameAs(records);
     }
 
@@ -344,7 +349,7 @@ class InBandKeyManagerTest {
     @Test
     void nullRecordValuesShouldNotBeModifiedAtEncryptTime() {
         InMemoryKms kms = getInMemoryKms();
-        var km = createKeyManager(kms, 500_000);
+        var encryptionManager = createEncryptionManager(kms, 500_000);
 
         var kekId = kms.generateKey();
 
@@ -352,7 +357,7 @@ class InBandKeyManagerTest {
 
         List<Record> encrypted = new ArrayList<>();
         List<Record> initial = List.of(record);
-        assertThat(doEncrypt(km, "topic", 1, new EncryptionScheme<>(kekId, EnumSet.of(RecordField.RECORD_VALUE)), initial, encrypted))
+        assertThat(doEncrypt(encryptionManager, "topic", 1, new EncryptionScheme<>(kekId, EnumSet.of(RecordField.RECORD_VALUE)), initial, encrypted))
                 .isCompleted();
         assertEquals(1, encrypted.size());
         assertFalse(encrypted.get(0).hasValue());
@@ -364,7 +369,7 @@ class InBandKeyManagerTest {
     @Test
     void nullRecordValuesAreIncompatibleWithHeaderEncryption() {
         InMemoryKms kms = getInMemoryKms();
-        var km = createKeyManager(kms, 500_000);
+        var encryptionManager = createEncryptionManager(kms, 500_000);
 
         var kekId = kms.generateKey();
 
@@ -374,7 +379,7 @@ class InBandKeyManagerTest {
         List<Record> encrypted = new ArrayList<>();
         List<Record> initial = List.of(record);
         String expectedMessage = "encrypting headers prohibited when original record value null, we must preserve the null for tombstoning";
-        assertThat(doEncrypt(km, "topic", 1, new EncryptionScheme<>(kekId, EnumSet.of(RecordField.RECORD_HEADER_VALUES)), initial, encrypted))
+        assertThat(doEncrypt(encryptionManager, "topic", 1, new EncryptionScheme<>(kekId, EnumSet.of(RecordField.RECORD_HEADER_VALUES)), initial, encrypted))
                 .failsWithin(Duration.ofSeconds(5)).withThrowableThat()
                 .withMessageContaining(expectedMessage);
     }
@@ -382,13 +387,13 @@ class InBandKeyManagerTest {
     @Test
     void shouldTolerateEncryptingEmptyBatch() {
         InMemoryKms kms = getInMemoryKms();
-        var km = createKeyManager(kms, 500_000);
+        var encryptionManager = createEncryptionManager(kms, 500_000);
 
         var kekId = kms.generateKey();
 
         List<Record> encrypted = new ArrayList<>();
         List<Record> initial = List.of();
-        assertThat(doEncrypt(km, "topic", 1, new EncryptionScheme<>(kekId, EnumSet.of(RecordField.RECORD_VALUE)), initial, encrypted))
+        assertThat(doEncrypt(encryptionManager, "topic", 1, new EncryptionScheme<>(kekId, EnumSet.of(RecordField.RECORD_VALUE)), initial, encrypted))
                 .isCompleted();
 
         assertEquals(0, encrypted.size());
@@ -397,10 +402,10 @@ class InBandKeyManagerTest {
     @Test
     void shouldTolerateEncryptingSingleBatchMemoryRecordsWithNoRecords() {
         InMemoryKms kms = getInMemoryKms();
-        var km = createKeyManager(kms, 500_000);
+        var encryptionManager = createEncryptionManager(kms, 500_000);
         EncryptionScheme<UUID> scheme = createScheme(kms);
         MemoryRecords records = RecordTestUtils.memoryRecordsWithAllRecordsRemoved();
-        assertThat(encrypt(km, scheme, records)).succeedsWithin(Duration.ZERO).isSameAs(records);
+        assertThat(encrypt(encryptionManager, scheme, records)).succeedsWithin(Duration.ZERO).isSameAs(records);
     }
 
     @Test
@@ -408,7 +413,7 @@ class InBandKeyManagerTest {
         InMemoryKms kms = getInMemoryKms();
         var kekId = kms.generateKey();
         // configure 1 encryption per dek but then try to encrypt 2 records, will destroy and retry
-        var km = createKeyManager(kms, 1);
+        var encryptionManager = createEncryptionManager(kms, 1);
 
         var value = ByteBuffer.wrap(new byte[]{ 1, 2, 3 });
         var value2 = ByteBuffer.wrap(new byte[]{ 4, 5, 6 });
@@ -417,7 +422,7 @@ class InBandKeyManagerTest {
 
         List<Record> encrypted = new ArrayList<>();
         List<Record> initial = List.of(record, record2);
-        CompletionStage<Void> encrypt = doEncrypt(km, "topic", 1, new EncryptionScheme<>(kekId, EnumSet.of(RecordField.RECORD_VALUE)),
+        CompletionStage<Void> encrypt = doEncrypt(encryptionManager, "topic", 1, new EncryptionScheme<>(kekId, EnumSet.of(RecordField.RECORD_VALUE)),
                 initial,
                 encrypted);
         assertThat(encrypt).failsWithin(Duration.ofSeconds(5)).withThrowableThat()
@@ -430,7 +435,7 @@ class InBandKeyManagerTest {
         var kekId = kms.generateKey();
         InMemoryKms spyKms = Mockito.spy(kms);
         when(spyKms.generateDekPair(kekId)).thenReturn(CompletableFuture.failedFuture(new EncryptorCreationException("failed to create that DEK")));
-        var km = createKeyManager(spyKms, 500000);
+        var encryptionManager = createEncryptionManager(spyKms, 500_000);
 
         var value = ByteBuffer.wrap(new byte[]{ 1, 2, 3 });
         var value2 = ByteBuffer.wrap(new byte[]{ 4, 5, 6 });
@@ -439,7 +444,7 @@ class InBandKeyManagerTest {
 
         List<Record> encrypted = new ArrayList<>();
         List<Record> initial = List.of(record, record2);
-        CompletionStage<Void> encrypt = doEncrypt(km, "topic", 1, new EncryptionScheme<>(kekId, EnumSet.of(RecordField.RECORD_VALUE)),
+        CompletionStage<Void> encrypt = doEncrypt(encryptionManager, "topic", 1, new EncryptionScheme<>(kekId, EnumSet.of(RecordField.RECORD_VALUE)),
                 initial,
                 encrypted);
         assertThat(encrypt).failsWithin(Duration.ofSeconds(5)).withThrowableThat().withMessageContaining("failed to create that DEK");
@@ -452,7 +457,8 @@ class InBandKeyManagerTest {
         InMemoryKms spyKms = Mockito.spy(kms);
         doReturn(CompletableFuture.failedFuture(new KmsException("failed to create that DEK"))).when(spyKms).decryptEdek(any());
 
-        var km = createKeyManager(spyKms, 50000);
+        var encryptionManager = createEncryptionManager(spyKms, 500_000);
+        var decryptionManager = createDecryptionManager(spyKms);
 
         var value = ByteBuffer.wrap(new byte[]{ 1, 2, 3 });
         var value2 = ByteBuffer.wrap(new byte[]{ 4, 5, 6 });
@@ -461,13 +467,13 @@ class InBandKeyManagerTest {
 
         List<Record> encrypted = new ArrayList<>();
         List<Record> initial = List.of(record, record2);
-        CompletionStage<Void> encrypt = doEncrypt(km, "topic", 1, new EncryptionScheme<>(kekId, EnumSet.of(RecordField.RECORD_VALUE)),
+        CompletionStage<Void> encrypt = doEncrypt(encryptionManager, "topic", 1, new EncryptionScheme<>(kekId, EnumSet.of(RecordField.RECORD_VALUE)),
                 initial,
                 encrypted);
         assertThat(encrypt).succeedsWithin(Duration.ofSeconds(5));
 
         List<Record> decrypted = new ArrayList<>();
-        CompletionStage<Void> decrypt = doDecrypt(km, "topic", 1, encrypted, decrypted);
+        CompletionStage<Void> decrypt = doDecrypt(decryptionManager, "topic", 1, encrypted, decrypted);
         assertThat(decrypt).failsWithin(Duration.ofSeconds(5)).withThrowableThat().withMessageContaining("failed to create that DEK");
     }
 
@@ -478,7 +484,7 @@ class InBandKeyManagerTest {
         InMemoryKms spyKms = Mockito.spy(kms);
         when(spyKms.generateDekPair(kekId)).thenReturn(CompletableFuture.failedFuture(new KmsException("failed to create that DEK")));
 
-        var km = createKeyManager(spyKms, 50000);
+        var encryptionManager = createEncryptionManager(spyKms, 50_000);
 
         var value = ByteBuffer.wrap(new byte[]{ 1, 2, 3 });
         var value2 = ByteBuffer.wrap(new byte[]{ 4, 5, 6 });
@@ -487,7 +493,7 @@ class InBandKeyManagerTest {
 
         List<Record> encrypted = new ArrayList<>();
         List<Record> initial = List.of(record, record2);
-        CompletionStage<Void> encrypt = doEncrypt(km, "topic", 1, new EncryptionScheme<>(kekId, EnumSet.of(RecordField.RECORD_VALUE)),
+        CompletionStage<Void> encrypt = doEncrypt(encryptionManager, "topic", 1, new EncryptionScheme<>(kekId, EnumSet.of(RecordField.RECORD_VALUE)),
                 initial,
                 encrypted);
         assertThat(encrypt).failsWithin(Duration.ofSeconds(5)).withThrowableThat().withMessageContaining("failed to create that DEK");
@@ -496,7 +502,7 @@ class InBandKeyManagerTest {
         when(spyKms.generateDekPair(kekId)).thenCallRealMethod();
 
         // when
-        CompletionStage<Void> encrypt2 = doEncrypt(km, "topic", 1, new EncryptionScheme<>(kekId, EnumSet.of(RecordField.RECORD_VALUE)),
+        CompletionStage<Void> encrypt2 = doEncrypt(encryptionManager, "topic", 1, new EncryptionScheme<>(kekId, EnumSet.of(RecordField.RECORD_VALUE)),
                 initial,
                 encrypted);
 
@@ -508,7 +514,8 @@ class InBandKeyManagerTest {
     @Test
     void shouldEncryptRecordValueForMultipleRecords() throws ExecutionException, InterruptedException, TimeoutException {
         InMemoryKms kms = getInMemoryKms();
-        var km = createKeyManager(kms, 500_000);
+        var encryptionManager = createEncryptionManager(kms, 500_000);
+        var decryptionManager = createDecryptionManager(kms);
 
         var kekId = kms.generateKey();
 
@@ -520,7 +527,7 @@ class InBandKeyManagerTest {
 
         List<Record> encrypted = new ArrayList<>();
         List<Record> initial = List.of(record, record2);
-        doEncrypt(km, "topic", 1, new EncryptionScheme<>(kekId, EnumSet.of(RecordField.RECORD_VALUE)),
+        doEncrypt(encryptionManager, "topic", 1, new EncryptionScheme<>(kekId, EnumSet.of(RecordField.RECORD_VALUE)),
                 initial,
                 encrypted)
                 .toCompletableFuture()
@@ -532,7 +539,7 @@ class InBandKeyManagerTest {
         // TODO add assertion on headers
 
         List<Record> decrypted = new ArrayList<>();
-        doDecrypt(km, "foo", 1, encrypted, decrypted).toCompletableFuture()
+        doDecrypt(decryptionManager, "foo", 1, encrypted, decrypted).toCompletableFuture()
                 .get(10, TimeUnit.SECONDS);
 
         assertEquals(initial, decrypted);
@@ -541,7 +548,7 @@ class InBandKeyManagerTest {
     @Test
     void shouldGenerateNewDekIfOldDekHasNoRemainingEncryptions() throws ExecutionException, InterruptedException, TimeoutException {
         InMemoryKms kms = getInMemoryKms();
-        var km = createKeyManager(kms, 2);
+        var encryptionManager = createEncryptionManager(kms, 2);
 
         var kekId = kms.generateKey();
 
@@ -553,7 +560,7 @@ class InBandKeyManagerTest {
 
         List<Record> encrypted = new ArrayList<>();
         List<Record> initial = List.of(record, record2);
-        doEncrypt(km, "topic", 1, new EncryptionScheme<>(kekId, EnumSet.of(RecordField.RECORD_VALUE)),
+        doEncrypt(encryptionManager, "topic", 1, new EncryptionScheme<>(kekId, EnumSet.of(RecordField.RECORD_VALUE)),
                 initial,
                 encrypted).toCompletableFuture().get(10, TimeUnit.SECONDS);
         record.value().rewind();
@@ -561,7 +568,7 @@ class InBandKeyManagerTest {
 
         // at this point we have encrypted 2 records with the manager set to maximum 2 encryptions per dek
 
-        doEncrypt(km, "topic", 1, new EncryptionScheme<>(kekId, EnumSet.of(RecordField.RECORD_VALUE)),
+        doEncrypt(encryptionManager, "topic", 1, new EncryptionScheme<>(kekId, EnumSet.of(RecordField.RECORD_VALUE)),
                 initial,
                 encrypted).toCompletableFuture().get(10, TimeUnit.SECONDS);
 
@@ -579,7 +586,7 @@ class InBandKeyManagerTest {
     @Test
     void shouldGenerateNewDekIfOldOneHasSomeRemainingEncryptionsButNotEnoughForWholeBatch() throws ExecutionException, InterruptedException, TimeoutException {
         InMemoryKms kms = getInMemoryKms();
-        var km = createKeyManager(kms, 3);
+        var encryptionManager = createEncryptionManager(kms, 3);
 
         var kekId = kms.generateKey();
 
@@ -591,7 +598,7 @@ class InBandKeyManagerTest {
 
         List<Record> encrypted = new ArrayList<>();
         List<Record> initial = List.of(record, record2);
-        doEncrypt(km, "topic", 1, new EncryptionScheme<>(kekId, EnumSet.of(RecordField.RECORD_VALUE)),
+        doEncrypt(encryptionManager, "topic", 1, new EncryptionScheme<>(kekId, EnumSet.of(RecordField.RECORD_VALUE)),
                 initial,
                 encrypted)
                 .toCompletableFuture()
@@ -601,7 +608,7 @@ class InBandKeyManagerTest {
 
         // at this point we have encrypted 2 records with the manager set to maximum 3 encryptions per dek, so we need a new dek to encrypt 2 more records
 
-        doEncrypt(km, "topic", 1, new EncryptionScheme<>(kekId, EnumSet.of(RecordField.RECORD_VALUE)),
+        doEncrypt(encryptionManager, "topic", 1, new EncryptionScheme<>(kekId, EnumSet.of(RecordField.RECORD_VALUE)),
                 initial,
                 encrypted).toCompletableFuture().get(10, TimeUnit.SECONDS);
 
@@ -618,7 +625,7 @@ class InBandKeyManagerTest {
     @Test
     void shouldUseSameDekForMultipleBatches() throws ExecutionException, InterruptedException, TimeoutException {
         InMemoryKms kms = getInMemoryKms();
-        var km = createKeyManager(kms, 4);
+        var encryptionManager = createEncryptionManager(kms, 4);
 
         var kekId = kms.generateKey();
 
@@ -630,7 +637,7 @@ class InBandKeyManagerTest {
 
         List<Record> encrypted = new ArrayList<>();
         List<Record> initial = List.of(record, record2);
-        doEncrypt(km, "topic", 1, new EncryptionScheme<>(kekId, EnumSet.of(RecordField.RECORD_VALUE)),
+        doEncrypt(encryptionManager, "topic", 1, new EncryptionScheme<>(kekId, EnumSet.of(RecordField.RECORD_VALUE)),
                 initial,
                 encrypted)
                 .toCompletableFuture()
@@ -640,7 +647,7 @@ class InBandKeyManagerTest {
 
         // at this point we have encrypted 2 records with the manager set to maximum 4 encryptions per dek, so we do not need a new dek to encrypt 2 more records
 
-        doEncrypt(km, "topic", 1, new EncryptionScheme<>(kekId, EnumSet.of(RecordField.RECORD_VALUE)),
+        doEncrypt(encryptionManager, "topic", 1, new EncryptionScheme<>(kekId, EnumSet.of(RecordField.RECORD_VALUE)),
                 initial,
                 encrypted).toCompletableFuture().get(10, TimeUnit.SECONDS);
 
@@ -656,7 +663,8 @@ class InBandKeyManagerTest {
     @Test
     void shouldEncryptRecordHeaders() {
         InMemoryKms kms = getInMemoryKms();
-        var km = createKeyManager(kms, 500_000);
+        var encryptionManager = createEncryptionManager(kms, 500_000);
+        var decryptionManager = createDecryptionManager(kms);
 
         var kekId = kms.generateKey();
 
@@ -666,7 +674,7 @@ class InBandKeyManagerTest {
 
         List<Record> encrypted = new ArrayList<>();
         List<Record> initial = List.of(record);
-        assertThat(doEncrypt(km, "topic", 1, new EncryptionScheme<>(kekId, EnumSet.of(RecordField.RECORD_VALUE, RecordField.RECORD_HEADER_VALUES)),
+        assertThat(doEncrypt(encryptionManager, "topic", 1, new EncryptionScheme<>(kekId, EnumSet.of(RecordField.RECORD_VALUE, RecordField.RECORD_HEADER_VALUES)),
                 initial,
                 encrypted))
                 .isCompleted();
@@ -676,7 +684,7 @@ class InBandKeyManagerTest {
         assertNotEquals(initial, encrypted);
 
         List<Record> decrypted = new ArrayList<>();
-        assertThat(doDecrypt(km, "topicFoo", 1, encrypted, decrypted))
+        assertThat(doDecrypt(decryptionManager, "topicFoo", 1, encrypted, decrypted))
                 .isCompleted();
 
         assertEquals(List.of(RecordTestUtils.record(value, new RecordHeader("headerFoo", new byte[]{ 4, 5, 6 }))), decrypted);
@@ -685,7 +693,8 @@ class InBandKeyManagerTest {
     @Test
     void shouldEncryptRecordHeadersForMultipleRecords() throws ExecutionException, InterruptedException, TimeoutException {
         InMemoryKms kms = getInMemoryKms();
-        var km = createKeyManager(kms, 500_000);
+        var encryptionManager = createEncryptionManager(kms, 500_000);
+        var decryptionManager = createDecryptionManager(kms);
 
         var kekId = kms.generateKey();
 
@@ -703,7 +712,7 @@ class InBandKeyManagerTest {
 
         List<Record> encrypted = new ArrayList<>();
         List<Record> initial = List.of(record, record2, record3);
-        doEncrypt(km, "topic", 1, new EncryptionScheme<>(kekId, EnumSet.of(RecordField.RECORD_VALUE, RecordField.RECORD_HEADER_VALUES)),
+        doEncrypt(encryptionManager, "topic", 1, new EncryptionScheme<>(kekId, EnumSet.of(RecordField.RECORD_VALUE, RecordField.RECORD_HEADER_VALUES)),
                 initial,
                 encrypted).toCompletableFuture().get(10, TimeUnit.SECONDS);
         value.rewind();
@@ -713,7 +722,7 @@ class InBandKeyManagerTest {
         assertNotEquals(initial, encrypted);
 
         List<Record> decrypted = new ArrayList<>();
-        assertThat(doDecrypt(km, "foo", 1, encrypted, decrypted)).isCompleted();
+        assertThat(doDecrypt(decryptionManager, "foo", 1, encrypted, decrypted)).isCompleted();
 
         assertEquals(List.of(RecordTestUtils.record(0L, value, new RecordHeader("foo", new byte[]{ 4, 5, 6 })),
                 RecordTestUtils.record(1L, value2, new RecordHeader("foo", new byte[]{ 10, 11, 12 })),
@@ -723,7 +732,8 @@ class InBandKeyManagerTest {
     @Test
     void shouldPropagateHeadersInClearWhenNotEncryptingHeaders() {
         InMemoryKms kms = getInMemoryKms();
-        var km = createKeyManager(kms, 500_000);
+        var encryptionManager = createEncryptionManager(kms, 500_000);
+        var decryptionManager = createDecryptionManager(kms);
 
         var kekId = kms.generateKey();
 
@@ -732,7 +742,7 @@ class InBandKeyManagerTest {
         var record = RecordTestUtils.record(ByteBuffer.wrap(value), header);
 
         List<Record> encrypted = new ArrayList<>();
-        assertThat(doEncrypt(km, "topic", 1, new EncryptionScheme<>(kekId, EnumSet.of(RecordField.RECORD_VALUE)), List.of(record), encrypted))
+        assertThat(doEncrypt(encryptionManager, "topic", 1, new EncryptionScheme<>(kekId, EnumSet.of(RecordField.RECORD_VALUE)), List.of(record), encrypted))
                 .isCompleted();
         assertThat(encrypted.iterator())
                 .toIterable()
@@ -743,7 +753,7 @@ class InBandKeyManagerTest {
                 .contains(header);
 
         List<Record> decrypted = new ArrayList<>();
-        assertThat(doDecrypt(km, "foo", 1, encrypted, decrypted))
+        assertThat(doDecrypt(decryptionManager, "foo", 1, encrypted, decrypted))
                 .isCompleted();
 
         assertThat(decrypted.iterator())
@@ -767,7 +777,8 @@ class InBandKeyManagerTest {
 
         var spyKms = Mockito.spy(kms);
 
-        var km = createKeyManager(spyKms, 50000);
+        var encryptionManager = createEncryptionManager(kms, 500_000);
+        var decryptionManager = createDecryptionManager(kms);
 
         byte[] rec1Bytes = { 1, 2, 3 };
         byte[] rec2Bytes = { 4, 5, 6 };
@@ -775,10 +786,10 @@ class InBandKeyManagerTest {
         var rec2 = RecordTestUtils.record(offsetB, ByteBuffer.wrap(rec2Bytes));
 
         List<Record> encrypted = new ArrayList<>();
-        var encryptStage = doEncrypt(km, topic, partition, new EncryptionScheme<>(kekId1, EnumSet.of(RecordField.RECORD_VALUE)),
+        var encryptStage = doEncrypt(encryptionManager, topic, partition, new EncryptionScheme<>(kekId1, EnumSet.of(RecordField.RECORD_VALUE)),
                 List.of(rec1),
                 encrypted)
-                .thenApply(u -> doEncrypt(km, topic, partition, new EncryptionScheme<>(kekId2, EnumSet.of(RecordField.RECORD_VALUE)),
+                .thenApply(u -> doEncrypt(encryptionManager, topic, partition, new EncryptionScheme<>(kekId2, EnumSet.of(RecordField.RECORD_VALUE)),
                         List.of(rec2),
                         encrypted));
         assertThat(encryptStage).isCompleted();
@@ -804,7 +815,7 @@ class InBandKeyManagerTest {
         }).when(spyKms).decryptEdek(argument.capture());
 
         List<Record> decrypted = new ArrayList<>();
-        var decryptStage = doDecrypt(km, topic, partition, encrypted, decrypted);
+        var decryptStage = doDecrypt(decryptionManager, topic, partition, encrypted, decrypted);
         assertThat(decryptStage).succeedsWithin(Duration.ofSeconds(1));
         assertThat(decrypted.iterator())
                 .toIterable()
@@ -824,7 +835,8 @@ class InBandKeyManagerTest {
         InMemoryKms kms = getInMemoryKms();
         var kekId = kms.generateKey();
 
-        var km = createKeyManager(kms, 50000);
+        var encryptionManager = createEncryptionManager(kms, 500_000);
+        var decryptionManager = createDecryptionManager(kms);
 
         byte[] rec1Bytes = { 1, 2, 3 };
         byte[] rec2Bytes = { 4, 5, 6 };
@@ -835,7 +847,7 @@ class InBandKeyManagerTest {
 
         // rec1 and rec3 will be encrypted.
         List<Record> encrypted = new ArrayList<>();
-        var encryptStage = doEncrypt(km, topic, partition, new EncryptionScheme<>(kekId, EnumSet.of(RecordField.RECORD_VALUE)),
+        var encryptStage = doEncrypt(encryptionManager, topic, partition, new EncryptionScheme<>(kekId, EnumSet.of(RecordField.RECORD_VALUE)),
                 List.of(rec1, rec3),
                 encrypted);
         assertThat(encryptStage).isCompleted();
@@ -846,7 +858,7 @@ class InBandKeyManagerTest {
         decryptInput.add(1, rec2);
 
         List<Record> received = new ArrayList<>();
-        var decryptStage = doDecrypt(km, topic, partition, decryptInput, received);
+        var decryptStage = doDecrypt(decryptionManager, topic, partition, decryptInput, received);
         assertThat(decryptStage).succeedsWithin(Duration.ofSeconds(1));
         assertThat(received.iterator())
                 .toIterable()
@@ -875,8 +887,8 @@ class InBandKeyManagerTest {
 
     @NonNull
     private static List<TestingDek> extractEdeks(List<Record> encrypted) {
-        List<TestingDek> deks = encrypted.stream()
-                .filter(testingRecord -> Stream.of(testingRecord.headers()).anyMatch(header -> header.key().equals(InBandKeyManager.ENCRYPTION_HEADER_NAME)))
+        return encrypted.stream()
+                .filter(testingRecord -> Stream.of(testingRecord.headers()).anyMatch(header -> header.key().equals(InBandDecryptionManager.ENCRYPTION_HEADER_NAME)))
                 .map(testingRecord -> {
                     ByteBuffer wrapper = testingRecord.value();
                     var edekLength = ByteUtils.readUnsignedVarint(wrapper);
@@ -885,12 +897,16 @@ class InBandKeyManagerTest {
                     return new TestingDek(edekBytes);
                 })
                 .toList();
-        return deks;
     }
 
     @NonNull
-    private static InBandKeyManager<UUID, InMemoryEdek> createKeyManager(InMemoryKms kms, int maxEncryptionsPerDek) {
-        return new InBandKeyManager<>(kms, BufferPool.allocating(), maxEncryptionsPerDek);
+    private static InBandDecryptionManager<UUID, InMemoryEdek> createDecryptionManager(InMemoryKms kms) {
+        return new InBandDecryptionManager<>(kms);
+    }
+
+    @NonNull
+    private static InBandEncryptionManager<UUID, InMemoryEdek> createEncryptionManager(InMemoryKms kms, int maxEncryptionsPerDek) {
+        return new InBandEncryptionManager<>(kms, BufferPool.allocating(), maxEncryptionsPerDek, 5_000_000_000L);
     }
 
     @NonNull
@@ -900,7 +916,7 @@ class InBandKeyManagerTest {
     }
 
     @NonNull
-    private static CompletionStage<MemoryRecords> decrypt(InBandKeyManager<UUID, InMemoryEdek> km, MemoryRecords encrypted) {
+    private static CompletionStage<MemoryRecords> decrypt(InBandDecryptionManager<UUID, InMemoryEdek> km, MemoryRecords encrypted) {
         return km.decrypt("topic", 1, encrypted, ByteBufferOutputStream::new);
     }
 
@@ -911,7 +927,7 @@ class InBandKeyManagerTest {
     }
 
     @NonNull
-    private static CompletionStage<MemoryRecords> encrypt(InBandKeyManager<UUID, InMemoryEdek> km, EncryptionScheme<UUID> scheme, MemoryRecords records) {
+    private static CompletionStage<MemoryRecords> encrypt(InBandEncryptionManager<UUID, InMemoryEdek> km, EncryptionScheme<UUID> scheme, MemoryRecords records) {
         return km.encrypt("topic", 1, scheme, records, ByteBufferOutputStream::new);
     }
